#ROBOMERGE-OWNER: jeanmichel.dignard
#ROBOMERGE-AUTHOR: josh.adams
#ROBOMERGE-SOURCE: CL 12152458 via CL 12152479
#ROBOMERGE-BOT: TOOLS (Main -> Dev-Tools-Staging) (v661-12148976)

[CL 12181877 by josh adams in Dev-Tools-Staging branch]

+++ Engine/Source/Runtime/Core/Private/Misc/FileHelper.cpp
@@ -256,7 +256,7 @@ bool FFileHelper::LoadFileToStringArrayWithPredicate(TArray<FString>& Result, co
 		}
 
 		FString Buffer;
-		BufferToString(Buffer, RawBuffer.GetData(), RawBuffer.Num());
+		BufferToString(Buffer, RawBuffer.GetData(), (int32)RawBuffer.Num());
 
 		for (const TCHAR* Pos = *Buffer; *Pos != 0; )
 		{
